Telegrambot/requirements.txt
Flask
python-telegram-bot==13.0
python-dotenv
pyngrok
